+++ system/src/sheets/ACActorSheet.mjs
@@ -5,7 +5,8 @@ import { onManageActiveEffect, prepareActiveEffectCategories } from "../helpers/
  * Extend the basic ActorSheet with some very simple modifications
  * @extends {ActorSheet}
  */
-export default class ACActorSheet extends ActorSheet {
+export default class ACActorSheet
+	extends foundry.appv1.sheets.ActorSheet {
 
 	/** @override */
 	static get defaultOptions() {