+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Chores
 - [#159] Merge new Italian translation updates from Crowdin
 - [#165] Accessing the global "CombatTracker" which is now namespaced under foundry.applications.sidebar.tabs.CombatTracker
+- [#166] Accessing the global "ActorSheet" which is now namespaced under foundry.appv1.sheets.ActorSheet
 
 ---